Move helper to db package
